Updated spring package to be com.github.dozermapper.spring
