+++ spring/src/main/java/com/github/dozermapper/spring/DozerBeanMapperFactoryBean.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.dozer.spring;
+package com.github.dozermapper.spring;
 
 import java.io.IOException;
 import java.net.URL;
