+++ spring/src/main/java/com/github/dozermapper/spring/MapperDefinitionParser.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.dozer.spring;
+package com.github.dozermapper.spring;
 
 import org.w3c.dom.Element;
