+++ spring/src/main/resources/META-INF/spring.handlers
@@ -1,1 +1,1 @@
-http\://dozermapper.github.io/schema/dozer-spring=org.dozer.spring.DozerNamespaceHandler
+http\://dozermapper.github.io/schema/dozer-spring=com.github.dozermapper.spring.DozerNamespaceHandler
