+++ spring/src/test/java/com/github/dozermapper/spring/functional_tests/support/EventTestListener.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.dozer.spring.functional_tests.support;
+package com.github.dozermapper.spring.functional_tests.support;
 
 import java.util.concurrent.atomic.AtomicInteger;
