+++ spring/src/test/resources/mappings/mappingSpring.xml
@@ -24,9 +24,9 @@
         <stop-on-errors>true</stop-on-errors>
         <wildcard>true</wildcard>
         <custom-converters>
-            <converter type="org.dozer.spring.functional_tests.support.InjectedCustomConverter">
-                <class-a>org.dozer.spring.vo.Source</class-a>
-                <class-b>org.dozer.spring.vo.Destination</class-b>
+            <converter type="com.github.dozermapper.spring.functional_tests.support.InjectedCustomConverter">
+                <class-a>com.github.dozermapper.spring.vo.Source</class-a>
+                <class-b>com.github.dozermapper.spring.vo.Destination</class-b>
             </converter>
         </custom-converters>
     </configuration>
