+++ spring/src/test/resources/spring-bean-factory.xml
@@ -31,12 +31,12 @@
         <property name="factories">
             <map>
                 <!-- the key matches the name of the factory in the dozerBeanMapping.xml file -->
-                <entry key="org.dozer.spring.functional_tests.support.SampleCustomBeanFactory" value-ref="customBeanFactory"/>
+                <entry key="com.github.dozermapper.spring.functional_tests.support.SampleCustomBeanFactory" value-ref="customBeanFactory"/>
                 <!-- more factories can be supplied with additional entry's -->
             </map>
         </property>
     </bean>
 
-    <bean id="customBeanFactory" class="org.dozer.spring.functional_tests.support.SampleCustomBeanFactory" scope="singleton"/>
+    <bean id="customBeanFactory" class="com.github.dozermapper.spring.functional_tests.support.SampleCustomBeanFactory" scope="singleton"/>
 
 </beans>
