+++ spring/src/test/resources/spring-bean-mapping-builder.xml
@@ -19,10 +19,10 @@
 <!DOCTYPE beans PUBLIC "-//SPRING//DTD BEAN//EN"
         "http://www.springframework.org/dtd/spring-beans.dtd">
 <beans>
-    <bean id="factoryWithMappingBuilder" class="org.dozer.spring.DozerBeanMapperFactoryBean">
+    <bean id="factoryWithMappingBuilder" class="com.github.dozermapper.spring.DozerBeanMapperFactoryBean">
         <property name="mappingBuilders">
             <list>
-                <bean class="org.dozer.spring.functional_tests.support.SampleBeanMappingBuilder"/>
+                <bean class="com.github.dozermapper.spring.functional_tests.support.SampleBeanMappingBuilder"/>
             </list>
         </property>
     </bean>
