+++ spring/src/test/resources/spring-custom-converter.xml
@@ -31,7 +31,7 @@
         </property>
     </bean>
 
-    <bean id="converter" class="org.dozer.spring.functional_tests.support.InjectedCustomConverter">
+    <bean id="converter" class="com.github.dozermapper.spring.functional_tests.support.InjectedCustomConverter">
         <property name="parameter" value="injectedName"/>
     </bean>
