+++ spring/src/test/resources/spring-event-listener.xml
@@ -30,6 +30,6 @@
         </property>
     </bean>
 
-    <bean id="eventTestListener" class="org.dozer.spring.functional_tests.support.EventTestListener" scope="singleton"/>
+    <bean id="eventTestListener" class="com.github.dozermapper.spring.functional_tests.support.EventTestListener" scope="singleton"/>
 
 </beans>
