+++ spring/src/test/resources/springNameSpace.xml
@@ -25,7 +25,7 @@
 
     <dozer:mapper id="beanMapper"/>
 
-    <bean class="org.dozer.spring.functional_tests.support.ReferencingBean"/>
+    <bean class="com.github.dozermapper.spring.functional_tests.support.ReferencingBean"/>
 
 </beans>
